# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.DebtTransactionDirection:
    enum:
    - INCOMING
    - OUTGOING
    type: string
    x-enum-varnames:
    - IncomingDebt
    - OutgoingDebt
  dto.DebtTransactionResponse:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      currency:
        type: string
      description:
        type: string
      groupExpenseId:
        type: string
      id:
        type: string
      isFromExpense:
        type: boolean
      profile:
        $ref: '#/definitions/dto.SimpleProfile'
      transferMethod:
        type: string
      type:
        description: '"LENT" or "BORROWED"'
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.FriendBalance:
    properties:
      netBalance:
        type: number
      totalBorrowedFromFriend:
        type: number
      totalLentToFriend:
        type: number
      transactionHistory:
        items:
          $ref: '#/definitions/dto.FriendTransactionItem'
        type: array
    type: object
...
  dto.FriendTransactionItem:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      transferMethod:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.InternalLoginRequest:
    properties:
      email:
        minLength: 3
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  dto.NewDebtTransactionRequest:
    properties:
      amount:
        type: number
      currency:
        type: string
      description:
        type: string
      direction:
        allOf:
        - $ref: '#/definitions/dto.DebtTransactionDirection'
        enum:
        - INCOMING
        - OUTGOING
      friendProfileId:
        type: string
      transferMethodId:
        type: string
    required:
    - amount
    - friendProfileId
    - transferMethodId
    type: object
...
  dto.RefreshTokenRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  dto.ResetPasswordRequest:
    properties:
      password:
        type: string
      passwordConfirmation:
        type: string
      token:
        minLength: 3
        type: string
    required:
    - password
    - passwordConfirmation
    - token
    type: object
  dto.SendPasswordResetRequest:
    properties:
      email:
        minLength: 3
        type: string
    required:
    - email
    type: object
  dto.SimpleProfile:
    properties:
      avatar:
        type: string
      id:
        type: string
      isUser:
        type: boolean
      name:
        type: string
    type: object
...
  dto.TokenResponse:
    properties:
      refreshToken:
        type: string
      token:
        type: string
      type:
        type: string
    type: object
...
  map_string_dto.FriendBalance:
    additionalProperties:
      $ref: '#/definitions/dto.FriendBalance'
    type: object
...
  response.JSONResponse-array_dto_DebtTransactionResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.DebtTransactionResponse'
        type: array
      errors:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
...
  response.JSONResponse-dto_DebtTransactionResponse:
    properties:
      data:
        $ref: '#/definitions/dto.DebtTransactionResponse'
      errors:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
...
  response.JSONResponse-dto_TokenResponse:
    properties:
      data:
        $ref: '#/definitions/dto.TokenResponse'
      errors:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
  response.JSONResponse-map_string_dto_FriendBalance:
    properties:
      data:
        $ref: '#/definitions/map_string_dto.FriendBalance'
      errors:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
...
  response.Pagination:
    properties:
      currentPage:
        type: integer
      hasNextPage:
        type: boolean
      hasPrevPage:
        type: boolean
      totalData:
        type: integer
      totalPages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Cashback backend API
  title: Cashback API
  version: "1.0"
paths:
  /auth/{provider}:
    get:
      parameters:
      - description: OAuth provider (e.g. google)
        in: path
        name: provider
        required: true
        type: string
      responses:
        "307":
          description: Temporary Redirect
      summary: Initiate OAuth2 login
      tags:
      - auth
  /auth/{provider}/callback:
    get:
      parameters:
      - description: OAuth provider (e.g. google)
        in: path
        name: provider
        required: true
        type: string
      - description: Authorization code from provider
        in: query
        name: code
        required: true
        type: string
      - description: State token
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-dto_TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Handle OAuth2 provider callback
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.InternalLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-dto_TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login with email and password
      tags:
      - auth
  /auth/logout:
    delete:
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Logout current session
      tags:
      - auth
  /auth/password-reset:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendPasswordResetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Send password reset email
      tags:
      - auth
  /auth/refresh:
    put:
      consumes:
      - application/json
      parameters:
      - description: Refresh token payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-dto_TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Refresh access token
      tags:
      - auth
  /auth/reset-password:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Reset password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Reset password using token
      tags:
      - auth
  /auth/verify-registration:
    get:
      parameters:
      - description: Verification token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-dto_TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Verify email registration
      tags:
      - auth
  /debts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-array_dto_DebtTransactionResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all debt transactions
      tags:
      - debts
    post:
      consumes:
      - application/json
      parameters:
      - description: New debt transaction payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.NewDebtTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.JSONResponse-dto_DebtTransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Record a new debt transaction
      tags:
      - debts
  /debts/recent:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-array_dto_DebtTransactionResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get recent debt transactions
      tags:
      - debts
  /debts/summary:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-map_string_dto_FriendBalance'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get debt transaction summary
      tags:
      - debts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"